Fix uninitialized memory use in TensorFlow LSTM integration.
`OpKernelContext::forward_input_or_allocate_temp` can yield a memory
region that either has the contents of the input tensor or
unintialized content. Unfortunately, there's no return code or simple
indicator to describe what's in the memory region you just got back.

Instead, we just take the const input tensor and const_cast the
const-ness away. While it's not super safe, it is fast and effective.
We only play this trick on the forward activations cache tensor (`v`)
since we know no other op in the graph needs to interact with it.