Add LayerNormLSTMCell, a CPU-compatible version of LayerNormLSTM.
This class, LayerNormLSTMCell, is parameter-compatible with
LayerNormLTSM. It's implemented using the TF Python API so it can
run on CPUs in addition to other accelerators. Another advantage
is that LayerNormLSTMCell is an instance of RNNCell, which means
it's not fused over time and can be used in e.g. autoregressive
models.

Note that LayerNormLSTMCell is not intended for training. In
particular, the kernel / recurrent kernel / bias initializers are
not customizable and the defaults are not very good.

Issue: #1