Fix unsafe use of cublasHandle_t across threads (TensorFlow).
Before this change, it was possible that two or more threads could
concurrently set the cuBLAS stream on a single cublasHandle_t.
Although the cuBLAS documentation states that the library is
thread-safe, we observed segfaults if proper care wasn't taken to
isolate the stream changes.

This change fixes the segfaults by creating per-thread cublasHandle_t
objects. It is assumed that TensorFlow will use a fixed thread pool
to issue ops.