Use default device for RNN layers in PyTorch.
When a Haste RNN layer is constructed in PyTorch, all the parameters
are placed on the GPU. It is then the caller's responsibility to
move the parameters to the CPU if desired. This behavior is
non-standard for PyTorch layers and assumes the availability of a
CUDA-enabled GPU, which is not always the case.

This change switches the behavior to use the default device when a
Haste RNN layer is constructed. The caller can then choose to move
the parameters to any other device if they wish.

BREAKING CHANGE:
  Haste layers in PyTorch no longer default to GPU. Callers must
  either set the default device to GPU before constructing a layer
  or move the layer to a GPU by calling `.cuda()`.