Add optional weight transforms to all layers (TensorFlow).
It's sometimes useful to modify weights before they're used by an
RNN layer. This change adds optional transform functions to weights
so that the caller has an opportunity to modify the weights before
they're used by the layer. Transforms are different from constraints
in that the on-disk weights don't change.