Fix segfaults and memory errors with CUDA 11
PyTorch no longer preserves the default cuBLAS pointer mode so we
need to set it explicitly before GEMM calls. This change in
behavior was the source of some illegal memory accesses.

In addition, building haste_pytorch with CppExtension instead of
CUDAExtension causes it to be linked against CUDA 10.2 instead of
CUDA 11 even though the rest of PyTorch is linked against CUDA 11.
As a result, a cuBLAS context created by PyTorch and used by
haste_pytorch can occasionally fail (e.g. on a GEMM with an inner
dimension > 4096) with a segfault.

Issue: #30